AudioFlinger: Control volume using Port ID
Bug: 317212590
Test:

This CL migrates the volume management within AudioFlinger
from stream type to port ID.

It gives full power to AudioPolicy to compute the list of port
(so MmapThreads/Tracks) on which volume control is required.

It prevents from overwritting MUSIC stream type which is
the default for volume groups without associated stream type.

Test: build & play audio
Change-Id: I4c9e8bb45660c9ceffcc0f4029b0617f9795ab3c