audioflinger: Handle playback state transitions
Policy Manager may map tracks to Offload/ Directs thread even if
apps/ clients doesn't request for the same explicitly.
Per current design, the Offload thread, when user stops playback,
expects the state transition as STOPPING_1 --> STOPPING_2 --> STOPPED.
But as the track type is not Offload/ Direct, the state is directly set
from ACTIVE to STOPPED. This specific condition results in a spin loop
to keep writing data to the HAL even when frameReady=0.
Update OFFLOAD thread implementation to transition from ACTIVE to
STOPPED_1-->STOPPED_2-->STOPPED even if the track is not of type
OFFLOAD.

Test: atest AudioTrackTest
Change-Id: I5881e08ceb3e071fe16e06ff4256049d8189b74f