Camera: check permission without package name
For the system native clent, the permission query is performed with the
mistaching package name, which is "client.pid<%d>".
This leads to the situation where the appops service throws a security
exception and the permission query value falls back to the default mode
with misleading logs.

This change will remove package name from the permission query, so that
the appops service retrives the package name from the UID.

Bug: 325404079
Test: check log with the system native camera client like face hal.
Change-Id: I221ea0da1be27bf602751e0a0c174a25208230a3