Fix up mediaswcodec security policies.
Add missing include of crash_dump policies for arm/arm64/riscv64
so that updates to crash_dump/debuggerd happen automatically.
Remove the old section from the policy that duplicated the crash_dump
information.

Add an explicit section for AddressSanitizer syscalls to the files
missing them to match all of the other swcodec policy files.

For arm/arm64, the crash_dump policy is too restrictive. In those
cases include a version of the crash_dump policy that doesn't
include the mmap/mprotect rules.

Test: Ran debuggerd <PID_OF_media.swcodec> on arm device.
Test: Ran debuggerd <PID_OF_media.swcodec> on arm64 device.
Test: Ran debuggerd <PID_OF_media.swcodec> on riscv64 emulator.
Test: Ran debuggerd <PID_OF_media.swcodec> on x86 emulator.
Test: Ran debuggerd <PID_OF_media.swcodec> on x86_64 emulator.
Change-Id: I33704daa92751537a0d1f277055b2118d2d249e5