Move the AIDL manifest declaration to swcodec apex
The service is served from the swcodec apex, so it should be the one to
declare it.

Because we use mainline prebuilts for `next` that serve the HIDL swcodec
interface, we need to keep including the entry in mediaserver. It's ok
to ALSO be declared by mediaserver. Once all of the prebuilts have been
updated to include their own declarations, we can remove it from
mediaserver.

Those prebuilts do not serve AIDL, so they don't need to have the AIDL entry
added by mediaserver.

Test: launch_cvd
Bug: 279809333
Change-Id: I7cd8d7aab46ae44e0a8ff983d8bd04cdc1c7f874