Revert "audio: add feature flag for new headset profile and SCO ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1caa70b6-77db-46a7-8b9f-ae5733455e70). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida7a644d176ec372c3f2c7ccb121e6247cca09fe