Revert "audio: add feature flag for new headset profile and SCO ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccbfd0d2-0456-4572-96f6-33fd4501aabe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5ebf8c1c44a8d359cbda275ac93746a53d1e233