Revert "VP9: Synthesize CSD from frame header"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd1a781c-4b25-444d-930f-39fc7079924c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If32b8bb75ceb092b0f1a064c5b8a3974a909910d