Revert "VP9: Synthesize CSD from frame header"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f5277dd-3a5d-4a33-acbd-37782cb6f53a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifef6b211fe05d0198c48fb8ebbc5ac0397793e1c