Revert "swcodec check for any installed HIDL version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2920daf-92ff-4d3b-9cd1-dd6af7d4bfba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f3aade3ebcb1434e88cde23f7db30c1292a878c