Revert "swcodec check for any installed HIDL version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66c9d123-3247-46b7-98ed-90a7996ebdfd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie19fed2b07024d6c0d1b18061c02914676c671ea