Revert "swcodec check for any installed HIDL version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d903b248-94df-4610-8461-47441bbde617). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic29524e163a430c04662369fc23d5c52f3bdbf7e