Revert "swcodec check for any installed HIDL version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ead5b04-2624-4def-9dbb-8b52fc69f29b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd4953131feeaf670a3384821c01f2de9aed8d34