Revert "swcodec check for any installed HIDL version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/23248ae6-36fc-4a10-8e48-5c2ceec056f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f7a73519bc71a25728d58a5bbb7c3378f793ac0