[Bugfix]Solve the problem that the kCodecOperationMode value obtained by MediaMetricsService is always 1 (1 is Sync mode)
Cause of the problem: Before mediametrcs is uploaded to Service, MediaCodec will call the setState function to set the state to UNINITIALIZED. At this time, kFlagIsAsync will be cleared from mFlags; after that, flushMediametrics will be called to upload mediametrcs to Service, so that the Service will obtain the kCodecOperationMode value. Always 1.

Solution: Move the kCodecOperationMode setting to the "Configure" and "Setcallback" phases; because the App may use synchronous mode (only configure) or asynchronous mode (may configure first and then setcallback or first setcallback and then configure), so The kCodecOperationMode setting is added in the "Configure" and "Setcallback" stages to ensure that there will be no omissions.

Change-Id: I8dc03d22d9ac9404b6bf654f2b16d00ce890ec3e
Signed-off-by: linyaozong1 <linyaozong1@xiaomi.corp-partner.google.com>