WIP avoid possible deadlock in StagefrightMetadataRetriever
EXPERIMENTAL EXPERIMENTAL EXPERIMENTAL

set up to avoid a deadlock in mediametadataretriever where the service
side caches codec instances and, in a busy environment, fills all the
binders and starves a codec resource manager request.

This does two things:
(1) prints some stats for how often we cache/inhibit caching of the
codec reference
(2) provides a cap on how many codec instances we'll allow to be cached.

The upper limit on the caching is by setting property debug.stagefright.max_metadataretriever
unset or == 0 means unlimited.  else, a small positive integer.
My thoughts are that <=5 is appropriate.  working towards this value is
why i added the stats tracking mentioned above.

as a WIP, the cache stats are emitted to logcat whenever a codec is
cached or when it could be reused from the cache (if present...) This
should help us know how often it matters.

Downsides of limiting the cache appear to be the performance cost of
re-initializing a codec that we might have re-used; there does not
appear to be any correctness issue with completely disabling the cache.

THIS CODE NEEDS TO BE CLEANED UP BEFORE IT IS READY TO MERGE

Bug: 330840841
Test: CtsMediaMiscTestCases
Change-Id: I31d49ae18471c11caebabc795b2ebc633f2a5677