Revert "Refactored libstagefright_mediamuxer_fuzzer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/073d5010-32c1-4feb-9d72-86a5107746cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9fd2836ae564d0cc6c85e1715b9dea4d1b8bd509