Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73cc0ab1-0406-4358-acb2-65a99cbbfd83). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2f3d37796bf94accab2ed33a5d57c5fe661a8a9