Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb4cf10f-6850-42dd-84cd-d95090d6766a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04ae40371d7f8eeb4ce5ddd199a2e8fea2f7ebe7