av: Fix service interface availability
Don't block FWK, if IConfig is not defined.
Along, with default check for other modules as well, if those are available,
if any of default/bluetooth/r_submix/usb/bluetooth/stub is available then consider
IModule to be available.

This can be helpful during target bring ups, when drivers are not up, vendors want
to use stub configurations.

Change-Id: I3cadbae9cf1b777f641070683ea04bc4a1b42bc2