Revert "Use the global default cpp_std."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9e6f335-d37c-4f1f-9fda-707992199d87). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 346434347
Change-Id: I637ba5bb08fee510857affa384b6add6625618ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com